Responsive Design

+++ didi.php
@@ -11,20 +11,27 @@ $mysqli = $_SESSION["connection"];
   <title>Did I win?</title>
   <meta name="description" content="QR code lottery">
   <meta name="author" content="Pano">
+  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, minimum-scale=1, user-scalable=no">
   <link href='http://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,600,300,400' rel='stylesheet' type='text/css'>
   <script src="http://code.jquery.com/jquery-1.9.1.min.js"></script>
   <style>
   body{
     font-family: 'Open Sans', sans-serif;
-    line-height: 150%;
     font-size: 120%;
+    width:auto;
+    padding:10px;
+    max-width:100%;
+    /*line-height: 150%;
+    font-size: 120%;*/
+  }
+  #input,
+  #winners{
+  	width:70%;
   }
   input[type='text']{
     border: 1px solid rgb(150,150,150);
     -moz-border-radius: 5px;
     border-radius: 5px;
-    line-height:30px;
-    padding:2px;
   }
   </style>
   <script>
@@ -76,21 +83,18 @@ $mysqli = $_SESSION["connection"];
   //Last digit of IP
   $lastdigit = substr($ip,-1);
   //Win condition!
-  if ($lastdigit%3==0){
+  if ($lastdigit%3==0 && !alreadyWon()){
     //Check to see if IP already in database
-  	if(!alreadyWon()){
-	    echo "<div id=\"theinput\">Congratulations, you have won!<br> If you'd like to brag about it, enter your name here and press enter: <br />";
-	    echo '<input type="text" id="name" /></div>';
-  	}
-	else{
-		echo "Don't try to cheat, I know you've already won.";
-	}
-
-    
-  }//If they didn't win
-  else{
-  	echo "I'm sorry! You lost. Try going somewhere else.";
+    echo "<div id=\"theinput\">Congratulations, you won!<br> If you'd like to brag about it, enter your name here and press enter: <br />";
+    echo '<input type="text" id="name" /></div>';
   }
+  else if ($lastdigit%3==0){
+		echo "Don't try to cheat, I know you've already won.";
+  }
+  else{
+  	echo "I'm sorry! You lost. Try going somewhere else.";  	
+  }
+
   ?>
 <h3>The Winners</h3>
 <div id="winners"></div>
